Extract Android Studio update checker to its own module
