+++ features/android-studio-updates/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     kotlin("jvm")
-    application
     id("io.gitlab.arturbosch.detekt")
 }
 
@@ -9,20 +8,16 @@ version = "1.0"
 
 dependencies {
     implementation(projects.core.configuration)
+    implementation(projects.core.fetcher)
     implementation(projects.core.discord)
+    implementation(projects.core.scheduler)
     implementation(projects.core.logging)
-
-    implementation(projects.features.androidStudioUpdates)
 
     testImplementation(kotlin("test"))
 }
 
 tasks.getByName<Test>("test") {
     useJUnitPlatform()
-}
-
-application {
-    mainClass.set("dale.MainKt")
 }
 
 detekt {
