+++ features/android-studio-updates/src/main/kotlin/features/updates/androidstudio/AndroidStudioUpdateFeature.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dale.studio
+package features.updates.androidstudio
 
 import discord.DiscordBotScope
 import discord.channel.Channel
